+++ Exchange Rate Info/config.ini
@@ -1,3 +1,3 @@
 [DEFAULT]
-url1 = https://data.fixer.io/api/latest?access_key={}&format={}&base={}
-key = [enter key here]
+url = http://data.fixer.io/api/latest?access_key={}&format={}
+key = 73c294b7132a6dda18494b3faeac47ae
