+++ Weather/Pipfile
@@ -1,0 +1,15 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+requests = "*"
+bs4 = "*"
+python3-lxml = "*"
+lxml = "*"
+
+[requires]
+python_version = "3.7"
